refactor: GraphViewModel 주석 추가 및 BeginnerFunctionType 파일 이동

+++ app/src/main/java/com/devhjs/mathgraphstudy/presentation/graph/GraphViewModel.kt
@@ -12,6 +12,7 @@ import com.devhjs.mathgraphstudy.domain.model.math.PlaceholderNode
 import com.devhjs.mathgraphstudy.domain.model.math.PowerNode
 import com.devhjs.mathgraphstudy.domain.model.math.VariableNode
 import com.devhjs.mathgraphstudy.domain.model.math.VisualMathNode
+import com.devhjs.mathgraphstudy.domain.model.math.enums.BeginnerFunctionType
 import com.devhjs.mathgraphstudy.domain.model.math.enums.MathFunction
 import com.devhjs.mathgraphstudy.domain.model.math.enums.MathOperator
 import com.devhjs.mathgraphstudy.domain.model.math.toDisplayString
@@ -49,6 +50,10 @@ class GraphViewModel @Inject constructor(
     private var intersectionJob: Job? = null
     private var functionAddedCount = 0
 
+    /**
+     * 사용자의 UI 액션을 처리하고, 그에 따라 상태(State)를 업데이트합니다.
+     * 입력 처리, 모드 전환, 함수 추가/삭제, 뷰포트 변경 등을 담당합니다.
+     */
     fun onAction(action: GraphAction) {
         when (action) {
             is GraphAction.OnInput -> {
@@ -166,18 +171,16 @@ class GraphViewModel @Inject constructor(
         }
     }
 
+    /**
+     * 초보자 모드에서 사용자가 입력한 계수 값들(a, b, c, d)을 바탕으로
+     * 파싱 가능한 수식 문자열(예: "(1)*x + (2)")을 생성합니다.
+     */
     private fun constructBeginnerExpression(state: GraphState): String {
         val coeffs = state.beginnerCoefficients
         val a = coeffs["a"] ?: "1"
         val b = coeffs["b"] ?: "0"
         val c = coeffs["c"] ?: "0"
         val d = coeffs["d"] ?: "0"
-
-        // Helper to handle 1 and 0 logic if needed, but for now explicitly using values is safer
-        // We will assume "a", "b", "c" are numbers. 
-        // If empty, defaults are provided.
-        // We wrap coefficients in parentheses to be safe with negative numbers? e.g. -1*x vs (-1)*x
-        // Actually, let's keep it simple.
 
         return when (state.beginnerFunctionType) {
             BeginnerFunctionType.LINEAR -> "($a)*x + ($b)"
@@ -187,11 +190,14 @@ class GraphViewModel @Inject constructor(
         }
     }
 
+    /**
+     * 그래프 함수나 뷰포트가 변경될 때 교차점 계산을 요청합니다.
+     * 연속적인 변경(예: 드래그)에 대응하기 위해 약간의 지연(debounce)을 둡니다.
+     */
     private fun triggerIntersectionCalculation() {
         intersectionJob?.cancel()
         intersectionJob = viewModelScope.launch {
-            // Optional: Debounce for smoother viewport updates
-            delay(50) 
+            delay(50)
             val currentState = _state.value
             val intersections = calculateIntersections(currentState)
             _state.update { it.copy(intersections = intersections) }
@@ -200,28 +206,30 @@ class GraphViewModel @Inject constructor(
 
 
 
+    /**
+     * 현재 보이는 뷰포트 범위 내에서 활성화된 함수들 간의 교차점을
+     * 백그라운드 스레드에서 비동기로 계산합니다.
+     */
     private suspend fun calculateIntersections(state: GraphState): List<Offset> = withContext(Dispatchers.Default) {
-        // Assume screen width approx 1080px. Center is 540.
-        // We extend range slightly to catch intersections near edges
         val buffer = 5.0
         val startX = ((-540f - state.viewportOffsetX) / state.viewportScale) - buffer
         val endX = ((540f - state.viewportOffsetX) / state.viewportScale) + buffer
 
         val intersections = calculateIntersectionsUseCase(
             functions = state.functions,
-            rangeStart = startX.toDouble(),
-            rangeEnd = endX.toDouble()
+            rangeStart = startX,
+            rangeEnd = endX
         )
 
         intersections.map { (x, y) ->
             Offset(x.toFloat(), y.toFloat())
         }
     }
-    
-    // ... generateRandomColor ...
 
+    /**
+     * 그래프 선을 그릴 때 사용할 랜덤 색상을 ARGB Long 값으로 생성합니다.
+     */
     private fun generateRandomColor(): Long {
-        // ARGB format: 0xAARRGGBB
         val alpha = 0xFF
         val red = Random.nextInt(256)
         val green = Random.nextInt(256)
@@ -233,6 +241,10 @@ class GraphViewModel @Inject constructor(
                blue.toLong()
     }
 
+    /**
+     * UI 표현을 위한 노드 트리(VisualMathNode)를
+     * 실제 수학 계산을 위한 도메인 노드 트리(ExpressionNode)로 변환합니다.
+     */
     private fun VisualMathNode.toExpressionNode(): ExpressionNode {
         return when (this) {
             is NumberNode -> ExpressionNode.Constant(this.value.toDoubleOrNull() ?: 0.0)
@@ -271,6 +283,10 @@ class GraphViewModel @Inject constructor(
         }
     }
 
+    /**
+     * 계산용 도메인 노드 트리(ExpressionNode)를 UI 표현용 노드 트리(VisualMathNode)로 역변환합니다.
+     * 이 과정에서 0 더하기, 1 곱하기 등의 기본적인 식 간소화 로직이 적용됩니다.
+     */
     private fun ExpressionNode.toVisualNode(): VisualMathNode {
         return when (this) {
             is ExpressionNode.Constant -> {
@@ -283,7 +299,6 @@ class GraphViewModel @Inject constructor(
                 val leftViz = this.left.toVisualNode()
                 val rightViz = this.right.toVisualNode()
 
-                // Simplification Logic
                 val isLeftZero = leftViz is NumberNode && (leftViz.value == "0" || leftViz.value == "0.0")
                 val isLeftOne = leftViz is NumberNode && (leftViz.value == "1" || leftViz.value == "1.0")
                 val isRightZero = rightViz is NumberNode && (rightViz.value == "0" || rightViz.value == "0.0")
@@ -302,18 +317,14 @@ class GraphViewModel @Inject constructor(
                         else -> MathOperator.PLUS
                     }
 
-                    // 1. Addition with 0: x + 0 -> x
                     if (op == MathOperator.PLUS && isRightZero) return leftViz
                     if (op == MathOperator.PLUS && isLeftZero) return rightViz
 
-                    // 2. Subtraction with 0: x - 0 -> x
                     if (op == MathOperator.MINUS && isRightZero) return leftViz
 
-                    // 3. Multiplication by 1: 1 * x -> x
                     if (op == MathOperator.MULTIPLY && isLeftOne) return rightViz
                     if (op == MathOperator.MULTIPLY && leftViz is NumberNode && (leftViz.value == "1" || leftViz.value == "1.0")) return rightViz
 
-                    // 4. Multiplication by 0: 0 * x -> 0 (Be careful with formatting, usually 0)
                     if (op == MathOperator.MULTIPLY && isLeftZero) return NumberNode("0")
                     if (op == MathOperator.MULTIPLY && isRightZero) return NumberNode("0")
 
@@ -333,7 +344,7 @@ class GraphViewModel @Inject constructor(
                     "log" -> MathFunction.LOG
                     "ln" -> MathFunction.LN
                     "abs" -> MathFunction.ABS
-                    else -> MathFunction.SIN // Fallback
+                    else -> MathFunction.SIN
                 }
                 FunctionNode(
                     func = func,
